html

+++ src/main/java/com/exercisegenerator/domain/repository/ExerciseRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 
 public interface ExerciseRepository extends CrudRepository<Exercise, Long> {
-    List<Exercise> findByExamId(Integer id);
+    List<Exercise> findByExamId(Long id);
 
 }
